services/ia/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
google-generativeai
pydantic-settings